OeX_ES-57: Moves the logic about datetime serializetion to separete Mixin and use it.
